Refactor controller tests into per-source files
Replace the monolithic unit_test.go (463 lines), the stub suite_test.go
(envtest bootstrap with zero specs), and reconcile_full_test.go with
one test file per source file. Shared helpers (testScheme, newTestReconciler,
basePhare, ptr helpers, existence checkers) are consolidated in
testhelpers_test.go. Remove now-unused ginkgo/gomega direct dependencies.
